fix: a11y: Fix role inheritance for select options (#943)
Roles for select options were set on internal elements enclosed
inside the li element. This resulted li interfering with accessibility
tools while deducing the aria attributes for the enclosed items as all
accessibility were set on enclosed items.

This commit sets the role for enclosing li as presentation.

`role="presentation"` (or equivalently `role="none"`) is an ARIA
attribute that indicates an element is being used only for visual
presentation and should be ignored by assistive technologies like
screen readers. It effectively removes the semantic meaning of the
element while keeping its visual appearance and behavior.

This helps in deducing the aria attributes for the enclosed items
correctly.